[SL-ONLY] MATTER-6216: Refactor Oven App (#1057)
* initial consolidation into AppTask and manager file

* crtp interface for AppTask and OvenManager with common CustomerAppManager file

* Restyled by whitespace

* consolidate into AppTask

* add some missing briefs

* Apply suggestions from code review


* add comment for inherited class prototype

* add detail and remove unnecessary comment in AppTaskImpl file

* Restyled by whitespace

* update with new methods from release branch

* add a few missing brief

---------
