+++ examples/oven-app/silabs/include/CHIPProjectConfig.h
@@ -100,3 +100,6 @@
  *
  */
 #define CHIP_CONFIG_MRP_LOCAL_ACTIVE_RETRY_INTERVAL (2000_ms32)
+
+// Temporary setting to use CustomerAppTask for apps which are not upgraded to new architecture
+#define CHIP_SILABS_APP_USE_CUSTOMER_APP_TASK
