+++ examples/oven-app/silabs/src/OvenUI.cpp
@@ -20,7 +20,6 @@
 #include <string.h>
 
 #include "AppTask.h"
-#include "OvenManager.h"
 #include "OvenUI.h"
 #include "demo-ui-bitmaps.h"
 #include "dmd.h"
@@ -90,8 +89,7 @@ void OvenUI::DrawHeader(GLIB_Context_t * glibContext)
 
 void OvenUI::DrawCookTopState(GLIB_Context_t * glibContext)
 {
-    // Get CookTop state from OvenManager
-    bool cookTopState = OvenManager::GetInstance().GetCookTopState();
+    bool cookTopState = AppTask::GetAppTask().GetCookTopState();
 
     if (cookTopState)
     {
@@ -106,7 +104,7 @@ void OvenUI::DrawCookTopState(GLIB_Context_t * glibContext)
 void OvenUI::DrawOvenMode(GLIB_Context_t * glibContext)
 {
     // Get current oven mode from the Temperature Controlled Cabinet endpoint
-    uint8_t currentMode = OvenManager::GetInstance().GetCurrentOvenMode();
+    uint8_t currentMode = AppTask::GetAppTask().GetCurrentOvenMode();
 
     // LCD Display line max length is 16 characters, so allocate 17 bytes for the message (MODE: + mode text + null terminator)
     char message[LCD_MAX_LINE_LEN + 1];
